encapsulate data calls into react hooks
