+++ src/App.tsx
@@ -1,67 +1,26 @@
 import { useEffect, useState } from 'react'
-import { createClient, PostgrestError } from '@supabase/supabase-js'
 import './app.css'
-
-const supabase = createClient('https://lbfcegnaffgqvbllwimf.supabase.co', 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImxiZmNlZ25hZmZncXZibGx3aW1mIiwicm9sZSI6ImFub24iLCJpYXQiOjE3Mzk0MjMyMzMsImV4cCI6MjA1NDk5OTIzM30.71ztpxWbFxUJwyGxNv451treAOI8NkPeREwSLhmNHyc')
+import useFetchJellybeans from './hooks/useFetchJellybeans'
+import useInsertJellybean from './hooks/useInsertJellybean'
+import useUpdateJellybean from './hooks/useUpdateJellybean'
+import useDeleteJellybean from './hooks/useDeleteJellybean'
 
 type Jellybean = { id: string; flavor: string }
 
 function App() {
   const [jellybeans, setJellybeans] = useState<Jellybean[]>([])
-  const [newJellybeanFlavor, setNewJellybeanFlavor] = useState<string>('')
   const [newJellybeanId, setNewJellybeanId] = useState<string>('')
+  const [newJellybeanFlavor, setNewJellybeanFlavor] = useState<string>('')
+
+  const { fetchJellybeans } = useFetchJellybeans(setJellybeans)
+  const { insertJellybean } = useInsertJellybean(fetchJellybeans)
+  const { updateJellybean } = useUpdateJellybean(fetchJellybeans)
+  const { deleteJellybean } = useDeleteJellybean(fetchJellybeans)
 
   useEffect(() => {
     fetchJellybeans()
+    // eslint-disable-next-line react-hooks/exhaustive-deps -- only want this to fetch on first render
   }, [])
-
-  async function fetchJellybeans() {
-    const { data, error }: { data: Jellybean[] | null; error: PostgrestError | null } = await supabase.from('jellybeans').select()
-
-    if (error) {
-      console.error(`Error fetching jellybeans: ${error.message}`)
-      alert(`Failed to fetch jellybeans: ${error.message}`)
-      return
-    }
-
-    setJellybeans(data ?? [])
-  }
-
-  async function insertJellybean(flavor: string) {
-    const { error } = await supabase.from('jellybeans').insert({ flavor })
-
-    if (error) {
-      console.error(`Error inserting jellybean: ${error.message}`)
-      alert(`Failed to insert jellybean: ${error.message}`)
-      return
-    }
-
-    fetchJellybeans()
-  }
-
-  async function updateJellybean({ flavor, id }: { flavor: string; id: string }) {
-    const { error } = await supabase.from('jellybeans').update({ flavor }).eq('id', id)
-
-    if (error) {
-      console.error(`Error updating jellybean: ${error.message}`)
-      alert(`Failed to update jellybean: ${error.message}`)
-      return
-    }
-
-    fetchJellybeans()
-  }
-
-  async function deleteJellybean(id: string) {
-    const { error } = await supabase.from('jellybeans').delete().eq('id', id)
-
-    if (error) {
-      console.error(`Error deleting jellybean: ${error.message}`)
-      alert(`Failed to delete jellybean: ${error.message}`)
-      return
-    }
-
-    fetchJellybeans()
-  }
 
   return (
     <>
