+++ src/hooks/useSupabase.ts
@@ -1,0 +1,10 @@
+import { createClient } from '@supabase/supabase-js'
+
+const supabaseUrl = 'https://lbfcegnaffgqvbllwimf.supabase.co'
+const supabaseAnonKey = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImxiZmNlZ25hZmZncXZibGx3aW1mIiwicm9sZSI6ImFub24iLCJpYXQiOjE3Mzk0MjMyMzMsImV4cCI6MjA1NDk5OTIzM30.71ztpxWbFxUJwyGxNv451treAOI8NkPeREwSLhmNHyc'
+
+const supabase = createClient(supabaseUrl, supabaseAnonKey)
+
+export function useSupabase() {
+  return supabase
+}
